requirements.txt
discord.py=1.7.3
apscheduler=3.9.1